third_party/blink/web_tests/fast/forms/text/input-maxlength-ime-completed-expected.txt

There was a bug that users could input text longer than maxlength via IME. This test confirms it was fixed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getSelection().toString() is "ab"
PASS successfullyParsed is true

TEST COMPLETE
